UpdaterCommon: Move most of the programs here
