+++ Source/Core/UpdaterCommon/CMakeLists.txt
@@ -5,4 +5,5 @@ target_link_libraries(updatercommon PRIVATE
   uicommon
   mbedtls
   z
-  ed25519)
+  ed25519
+  cpp-optparse)
